vmwatcher: remove controller abstraction, it was making code harder to read
