--- go/src/koding/vmcleaner/controller.go
@@ -1,42 +1,0 @@
-package main
-
-import (
-	"errors"
-	"fmt"
-	"koding/db/models"
-
-	"github.com/koding/kite"
-	"github.com/koding/kodingemail"
-)
-
-type Action func(*models.User, int) error
-
-type Controller struct {
-	Kite  *kite.Client
-	Email kodingemail.Client
-}
-
-var (
-	templates = map[int]string{
-		1: "8239b8ac-032d-4f90-80de-a144519a945c",
-		2: "a987a9c0-f116-4e75-bf18-bff3c1c11b37",
-		3: "fedf1228-4624-4c23-8dea-e9391c5d1e98",
-	}
-)
-
-func (c *Controller) SendEmail(user *models.User, levelId int) error {
-	templateId, ok := templates[levelId]
-	if !ok {
-		return fmt.Errorf("template not found for level: %v", levelId)
-	}
-
-	if c.Email == nil {
-		return errors.New("email client is not initialized")
-	}
-
-	return c.Email.SendTemplateEmail(user.Email, templateId, nil)
-}
-
-func (c *Controller) DeleteVM(user *models.User, levelId int) error {
-	return nil
-}
